move playlist related func to playlists.go + fix linter
