+++ src/main/main.go
@@ -1,18 +1,13 @@
 package main
 
 import (
-	"encoding/json"
 	"explo/src/logging"
 	"explo/src/models"
 	"explo/src/web"
-	"io"
 	"log"
 	"log/slog"
-	"net/http"
 	"os"
-	"path/filepath"
-	"strings"
-
+	
 	"explo/src/client"
 	"explo/src/config"
 	"explo/src/discovery"
@@ -30,69 +25,6 @@ func initHttpClient() *util.HttpClient {
 	return util.NewHttp(util.HttpClientConfig{
 		Timeout: 10,
 	})
-}
-
-// writePlaylistCache downloads cover art and writes a tracklist JSON for the web UI.
-// added maps "CleanTitle|Artist" → true for tracks that made it into the playlist; nil means status unknown.
-func writePlaylistCache(cfgPath, playlist string, tracks []*models.Track, added map[string]bool) {
-	type cachedTrack struct {
-		Rank      int    `json:"rank"`
-		Title     string `json:"title"`
-		Artist    string `json:"artist"`
-		Release   string `json:"release"`
-		CoverURL  string `json:"coverUrl,omitempty"`
-		InLibrary *bool  `json:"inLibrary,omitempty"`
-	}
-	type cache struct {
-		Tracks []cachedTrack `json:"tracks"`
-	}
-
-	cfgDir := filepath.Dir(cfgPath)
-	coversDir := filepath.Join(cfgDir, "cache", "covers")
-	os.MkdirAll(coversDir, 0755)
-
-	ct := make([]cachedTrack, len(tracks))
-	for i, t := range tracks {
-		localCover := ""
-		if t.CoverURL != "" {
-			// Use the CAA release MBID (second-to-last path segment) as filename.
-			parts := strings.Split(strings.TrimRight(t.CoverURL, "/"), "/")
-			mbid := parts[len(parts)-2]
-			destPath := filepath.Join(coversDir, mbid+".jpg")
-			if _, err := os.Stat(destPath); os.IsNotExist(err) {
-				if resp, err := http.Get(t.CoverURL); err == nil { //nolint:noctx
-					defer resp.Body.Close()
-					if resp.StatusCode == http.StatusOK {
-						if data, err := io.ReadAll(resp.Body); err == nil {
-							os.WriteFile(destPath, data, 0644)
-						}
-					}
-				}
-			}
-			localCover = "/api/covers/" + mbid + ".jpg"
-		}
-		var inLibrary *bool
-		if added != nil {
-			v := added[t.CleanTitle+"|"+t.Artist]
-			inLibrary = &v
-		}
-		ct[i] = cachedTrack{
-			Rank:      i + 1,
-			Title:     t.CleanTitle,
-			Artist:    t.Artist,
-			Release:   t.Album,
-			CoverURL:  localCover,
-			InLibrary: inLibrary,
-		}
-	}
-
-	raw, err := json.Marshal(cache{Tracks: ct})
-	if err != nil {
-		return
-	}
-	cacheDir := filepath.Join(cfgDir, "cache")
-	os.MkdirAll(cacheDir, 0755)
-	os.WriteFile(filepath.Join(cacheDir, playlist+".json"), raw, 0644)
 }
 
 // Inits debug, gets playlist name, if needed, handles deprecation
